// Angular13/src/app/models/Ilan.ts
export class Ilan {
    IlanId?: number;
    IlanBaslik?: string;
    IlanAciklama?:string;
    IlanFiyat?: number;
    IlanKatId?: number;
    IlanKatAdi?: string;
    IlanOda?: string;
    IlanKatsayisi?: number;
    IlanAdres?: string;
    IlanM2?: number;
    IlanDurum?: string;
    IlanTarih?: string;
    IlanBulunankat?: number;
    IlanFoto?: string;
    IlanUyeId?: number;
    IlanUyeAdi?: string;
    IlanUyeSoyadi?: string;
    IlanUyeTelefon?: string;
}